requirements.txt
streamlit
pandas
requests
lxml
openpyxl
python-dateutil
pyinstaller
gspread
gspread-dataframe
oauth2client
psycopg2-binary

